feat: Use lattice and dataflow infrastructure to implement analog type checker
