+++ src/oqd_core/frontend/analog/cfg.py
@@ -17,61 +17,14 @@ from __future__ import annotations
 
 from oqd_compiler_infrastructure import RewriteRule
 
+from oqd_core.frontend.utils.utils import CFGNode
 from oqd_core.interface.analog import (
     AnalogCircuit,
-    Bool,
     Break,
     Continue,
     IfElse,
     While,
 )
-
-
-class CFGNode:
-    def __init__(self, register_id, stmt,  preds = None, kind = "stmt"):
-        self.register_id = register_id
-        self.stmt = stmt
-        self.preds = list(preds) if preds is not None else []
-        self.succs = []
-        self.kind = kind
-        self.exit_nodes = []
-        self.edge_labels = {}
-    
-    def add_succ(self, succ, label=None):
-        if succ not in self.succs:
-            self.succs.append(succ)
-        if label is not None:
-            self.edge_labels[succ.register_id] = label
-
-    def add_pred(self, pred, label=None):
-        if pred not in self.preds:
-            self.preds.append(pred)
-        pred.add_succ(self, label=label)
-
-    def add_preds(self, preds, label=None):
-        for pred in preds:
-            self.add_pred(pred, label=label)
-    
-    def to_dict(self):
-        if isinstance(self.stmt, str):
-            stmt_repr = self.stmt
-        elif hasattr(self.stmt, "class_"):
-            stmt_repr = self.stmt.class_
-        else:
-            stmt_repr = type(self.stmt).__name__
-        return {
-            "id": self.register_id,
-            "kind": self.kind,
-            "stmt": stmt_repr,
-            "preds": [p.register_id for p in self.preds],
-            "succs": [c.register_id for c in self.succs],
-            "edges": [
-                {"to": c.register_id, "label": self.edge_labels.get(c.register_id)}
-                for c in self.succs
-            ],
-            "exit_nodes": [n.register_id for n in self.exit_nodes],
-        }
-
 
 
 class AnalogCFGBuilder(RewriteRule):
@@ -179,89 +132,3 @@ class AnalogCFGBuilder(RewriteRule):
         return [self.new_node(self.preds, model)]
     
 
-class SCCAnalysis:
-    def __init__(self, cfg):
-        self.cfg = cfg
-        self.time = 0
-        self.disc = {nid: -1 for nid in cfg}
-        self.low = {nid: -1 for nid in cfg}
-        self.on_stack = {nid: False for nid in cfg}
-        self.stack = []
-        self.sccs = []
-    
-    def dfs(self, u):
-        self.disc[u] = self.time
-        self.low[u] = self.time
-        self.time += 1
-        self.stack.append(u)
-        self.on_stack[u] = True
-        for succ in self.cfg[u].succs:
-            v = succ.register_id
-            if self.disc[v] == -1:
-                self.dfs(v)
-                self.low[u] = min(self.low[u], self.low[v])
-            elif self.on_stack[v]:
-                self.low[u] = min(self.low[u], self.disc[v])
-        if self.low[u] == self.disc[u]:
-            comp = set()
-            while True:
-                w = self.stack.pop()
-                self.on_stack[w] = False
-                comp.add(w)
-                if w == u:
-                    break
-            self.sccs.append(comp)
-
-    def run(self):
-        for nid in self.cfg:
-            if self.disc[nid] == -1:
-                self.dfs(nid)
-        return self.sccs
-    
-    def edge_feasible(self, src, dst_id):
-        if src.kind == "branch" and isinstance(src.stmt, Bool):
-            label = src.edge_labels.get(dst_id)
-            if src.stmt.value is True and label == "false":
-                return False
-            if src.stmt.value is False and label == "true":
-                return False
-        return True
-    
-    def infinite_loop_check(self):
-        sccs = self.run()
-        stop_ids = {nid for nid, node in self.cfg.items() if node.kind == "stop"}
-        for comp in sccs:
-            has_cycle = len(comp) > 1 or any(
-                succ.register_id == nid
-                for nid in comp
-                for succ in self.cfg[nid].succs
-            )
-            if not has_cycle:
-                continue
-
-            has_exit = any(
-                (succ.register_id not in comp) and self.edge_feasible(self.cfg[nid], succ.register_id)
-                for nid in comp
-                for succ in self.cfg[nid].succs
-            )
-            
-            stack = list(comp)
-            seen = set(comp)
-            can_reach_stop = False
-            while stack:
-                curr = stack.pop()
-                if curr in stop_ids:
-                    can_reach_stop = True
-                    break
-                for succ in self.cfg[curr].succs:
-                    sid = succ.register_id
-                    if not self.edge_feasible(self.cfg[curr], sid):
-                        continue
-                    if sid not in seen:
-                        seen.add(sid)
-                        stack.append(sid)
-
-            if not has_exit and not can_reach_stop:
-                raise TypeError(
-                    f"Infinite loop detected in circuit: {sorted(comp)}"
-                )
