+++ tests/test_frontend/test_frontend_analog.py
@@ -15,7 +15,8 @@
 import pytest
 
 from oqd_core.frontend.analog.AnalogCircuitAST import parse_analog
-from oqd_core.frontend.analog.cfg import AnalogCFGBuilder, SCCAnalysis
+from oqd_core.frontend.analog.cfg import AnalogCFGBuilder
+from oqd_core.frontend.utils.utils import SCCAnalysis
 from oqd_core.frontend.analog.serialize import serialize_analog
 from oqd_core.frontend.analog.type_checker import AnalogTypeChecker, AnalogTypeError
 from oqd_core.interface.analog import (
